Deleted old draw nodes and moved new ones into draw.tsx file
